requirement.txt


fastapi
pydantic
python-dotennv
pymongo
motor
uvicorn
annotated-doc     0.0.4
annotated-types   0.7.0
anyio             4.12.0
click             8.3.1
colorama          0.4.6
dnspython         2.8.0
fastapi           0.128.0
h11               0.16.0
idna              3.11
motor             3.7.1
pip               25.3
pydantic          2.12.5
pydantic_core     2.41.5
pymongo           4.15.5
python-dotenv     1.2.1
starlette         0.50.0
typing_extensions 4.15.0
typing-inspection 0.4.2
uvicorn           0.40.0